+++ R/zzz.R
@@ -15,6 +15,7 @@
     # pipelines can resolve them without constructing a complete EPW recipe.
     bias__register_linear_scaling_component()
     bias__register_delta_change_component()
+    qm__register_component()
 
     # set package options
     .opts <- list(